fixed error with null data and long data instead of int. code cleaning

+++ ddprofiler/src/main/java/analysis/IntegerDataConsumer.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public interface IntegerDataConsumer extends DataConsumer {
 
-	public boolean feedIntegerData(List<Integer> records);
+	public boolean feedIntegerData(List<Long> records);
 
 }
